Seeding no longer handled by PointManager
Seeding (initial and boundary) brought back into the main program. This separates the task of keeping track of points from the task of managing whether they should be created, kept, or destroyed. Culling is still handled by PointManager but this is expected to change - possibly a DomainManager or similar is needed.