Converted codebase to 3D
The point, point manager, and domain manager classes now work in 3D. Data is still read in for only one level (assumed to apply throughout the domain) but a pressure thickness is assumed. No pressure velocity is currently applied, as the goal was to verify no change in behavior when allowing vertical information to be included. Also fixed a rather nasty bug in output processing.