Made point manager abstract
The point manager class is now abstract to allow for multiple different kinds of point manager. This means that we can have, for example, a dense point manager designed to represent all air, and an exhaust point manager which only seeks to follow air affected by an aircraft (although still incurring the mass inconsistency nightmare). To facilitate this, the history (diagnostic output) has been made more flexible and points can return different properties based on a string request. This is going to be very slow - may need to try adding something which allows retrieval functions to be defined at startup, rather than just asking with the same string each time - but will come to that later if profiling shows a need.