Reduced time wasting during seeding
Point seeding was taking up most of the simulation time when a large number of flight segments were read in. Flight segments are now a) not read in if they are outside the simulation scope, b) skipped during seeding if not for the correct day, and c) deleted once complete.