Removal of boutique properties and groundwork for trajectory output
All properties are now stored in the PropertyHistory dictionary, rather than there being special lists for a small number of key properties (i.e. x, y, pressure, age). As part of this, UID is now output as a double, and referring to locations using x and y has been generally retired in favour of referring to longitude and latitude. Some additional groundwork has also been performed to prepare for trajectory output, but this is not yet functional.