Also create a flightmatching file
A file is now output, if using PointManagerFlight, which can be used to find all the trajectories associated with one flight. Specifically, the file contains three columns: UID, flight, and seed time. Each combination of UID and seed time is unique, so this can be used to identify all relevant trajectories.